refactor: reorganize user info retrieval by moving getRunningUserInfoByTmbId to user/team module and updating imports across the codebase
(cherry picked from commit 1cf99ec6a6dd1dc7bf8f4fa2ca67010bff643245)
