+++ projects/app/src/pages/api/v1/chat/completions.ts
@@ -34,10 +34,8 @@ import {
   removeEmptyUserInput
 } from '@fastgpt/global/core/chat/utils';
 import { updateApiKeyUsage } from '@fastgpt/service/support/openapi/tools';
-import {
-  getUserChatInfoAndAuthTeamPoints,
-  getRunningUserInfoByTmbId
-} from '@fastgpt/service/support/permission/auth/team';
+import { getUserChatInfoAndAuthTeamPoints } from '@fastgpt/service/support/permission/auth/team';
+import { getRunningUserInfoByTmbId } from '@fastgpt/service/support/permission/user/team';
 import { AuthUserTypeEnum } from '@fastgpt/global/support/permission/constant';
 import { MongoApp } from '@fastgpt/service/core/app/schema';
 import { type AppSchema } from '@fastgpt/global/core/app/type';
